Multiple splash + homepage settings updates
Decoupled the registration splash from the homepage splash (so it doesn't try to access user data that doesn't exist yet), added the option to speed up the splash screen (ala base minisrv's FastSplash), and more cleanly explained the classic home being forced on pre-Etude clients.
Yes, the additional input field is a dirty hack because there's no way to tell the difference between a setting omitted from a POST because it's disabled and a setting omitted because it's simply unchecked.